+++ backend/src/main/java/com/utn/magtea/modeloanimal/ModeloAnimalPoolAporteInputDTO.java
@@ -2,8 +2,10 @@ package com.utn.magtea.modeloanimal;
 
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record ModeloAnimalPoolAporteInputDTO(
         @NotNull(message = "El tubo del pool es obligatorio") Long poolTuboId,
-        Double cantidadConsumida,
+        BigDecimal cantidadConsumida,
         Integer dia
 ) {}
